MetaScreen: introduce window-added, window-removed, window-monitor-changed and window-workspace-changed signals.
Window changes in Cinnamon's window-list applet currently
listen to window-added and window-removed for both monitor
and workspace changes.  We end up acting on a lot of redundant
events, and these signals make no distinction between a window
simply being moved to a different monitor or workspace, and one
being created or destroyed (unmanaged).

The old signals are kept as is, for compatibility, but the new
signals will let our window-list respond to specific events, in
an efficient way, instead of doing things like destroying a
window-list item and immediately re-creating it because its window
has moved to a new monitor.